+++ GoogleAdManager/CHANGELOG.md
@@ -1,5 +1,9 @@
 # Changelog
 
+## 11.13.0.1
+* Add support for [inline adaptive banners in MRECs](https://developers.applovin.com/en/max/ios/ad-formats/banner-and-mrec-ads/#inline-adaptive-banners-in-mrecs).
+* Remove support for rewarded interstitial ads.
+
 ## 11.13.0.0
 * Certified with GoogleAdManager SDK 11.13.0.